🎨(app-desk) add feature members
The feature teams is getting big, we extracted codes
related to members to a new feature members.
