+++ src/frontend/apps/desk/src/features/members/index.ts
@@ -1,0 +1,2 @@
+export * from './types';
+export * from './components/MemberGrid';
