+++ src/frontend/apps/desk/src/features/members/types.tsx
@@ -1,0 +1,20 @@
+import { User } from '@/features/auth/';
+
+export enum Role {
+  MEMBER = 'member',
+  ADMIN = 'administrator',
+  OWNER = 'owner',
+}
+
+export interface Access {
+  id: string;
+  role: Role;
+  user: User;
+  abilities: {
+    delete: boolean;
+    get: boolean;
+    patch: boolean;
+    put: boolean;
+    set_role_to: Role[];
+  };
+}
