+++ src/frontend/apps/desk/src/features/teams/components/index.ts
@@ -1,3 +1,2 @@
 export * from './Panel/Panel';
 export * from './TeamInfo';
-export * from './Member/MemberGrid';
